feat(neovim): add nerdtree plugin
This patch adds NERDTree to the list of plugins. I mainly use fuzzy
search with fzf and tpope's projectionist in order to navigate code
bases but I do like to rely on a tree hierarchy from time to time.

NERDTree works great for that purpose, lazy loading it as I rarely rely
on it.