chore(git): update editor from nvim binary to nvim.appimage
This patch updates git's default editor switching from the apt package
neovim to the neovim's app image. As the distribution package was kind
of left behind I switched to that release in order to use the latest
stable and performs manual rolling updates.

I will still have to performs manual updates for now, but watching the
repo for releases only should ease that process.